Most of UserService is coded, most is not quite working correctly
Continuing to perfect UserService before starting on TweetService. Considering removal of Hashtags as an Entity. Dropped support for database connections between set of Hashtags and set of Users within Tweets accounting for the set of those mentions (Users) and the set of hashtags used in the Tweet, now they are represented by String labels (Hashtag) and Strng usernames (Users)