Final update. Minor changes to UzerService, failed to write parsers before deadline.
Much is not implemented. Likes are not implemented at all. Hashtags cannot be detected and added. Mentions are also not detected and added. Tweets continue to behave badly. Error messages exist in an unusual way, packaged inside dto objects that are eventually returned by the controller. This was not intended, but implementing better error handling is also not possible before the deadline.